Move the delegation/scope-return engine to lib/delegation/engine.js
Verbatim move of wireProcessHandlers, handleScopeReturn (with
spawnResumedProcess), and handleDelegation out of server.js, plus the
end_delegation shim body extracted as handleEndDelegation so the glue
genuinely joins lib/delegation/. The engine is composed once in the root
via createDelegationEngine(deps): live process state by identity, the
kill-window transition machine (which stays in the root; the engine calls
in through the injected capability), the transcript composition, the
auto-resume budget, permission config, and the error family. Lib modules
are required directly; function declarations stay at column 0 so coverage
anchors and source scans address them unchanged.

Three deliberate edits inside the moved text, everything else
byte-identical (script-built with asserts):
- 4x `cwd: WORKSPACE` -> `cwd: config.getWorkspace()` (read at use time)
- 1x `_claudeAuthEvidence = true` -> injected noteClaudeAuthEvidence()
- the end_delegation body dedented into handleEndDelegation(msg, processes)

Class B retargets (enumerated): the kill-window arming-site count scan
splits (server.js keeps the chat shim's site, engine carries >= 3); the
loop-guard respawn scan and the per-line exited-break scan read the
engine file. All other pins valid unretargeted.
